final commit with floats instead of int

+++ median.py
@@ -4,7 +4,7 @@
 while True:
     try:
         print("Enter a list of numbers separated by commas: ")
-        numbers = [int(value) for value in input().split(",")]
+        numbers = [float(value) for value in input().split(",")]
     except ValueError:
         print("Some input could not be converted to a number!")
     else:
